# docs/swagger.yaml
basePath: /
definitions:
...
  http.ErrorResponse:
    properties:
      message:
        description: Error message
        type: string
    type: object
  http.StatusResponse:
    properties:
      status:
        description: Status message
        type: string
    type: object
...
  role.Role:
    properties:
      id:
        description: Role ID
        type: string
      name:
        description: Role name
        type: string
    type: object
...
  user.CreateUserInput:
    properties:
      firstname:
        description: First name
        type: string
      lastname:
        description: Last Name
        type: string
      password:
        description: Password
        type: string
      phone:
        description: Phone number
        type: string
      roleid:
        description: Role ID
        type: integer
    required:
    - firstname
    - lastname
    - password
    - phone
    type: object
...
  user.UpdateUserInput:
    properties:
      firstname:
        description: First name
        type: string
      id:
        description: User ID
        type: string
      lastname:
        description: Last Name
        type: string
      password:
        description: Password
        type: string
    type: object
  user.User:
    properties:
      firstname:
        description: First name
        type: string
      id:
        description: User ID
        type: string
      lastname:
        description: Last Name
        type: string
      password:
        description: Password
        type: string
      phone:
        description: Phone number
        type: string
      role:
        description: Role name
        type: string
      roleid:
        description: Role ID
        type: integer
      status:
        description: Users status
        type: string
    required:
    - firstname
    - lastname
    - password
    - phone
    type: object
info:
  contact:
    email: evgeniydammer@gmail.com
    name: Evgeniy Dammer
  description: marketplace API service
  title: marketplace API
  version: "1.0"
paths:
...
  /api/v1/users/:
    get:
      consumes:
      - application/json
      description: Get all users method.
      produces:
      - application/json
      responses:
        "200":
          description: User List
          schema:
            items:
              $ref: '#/definitions/user.User'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - Bearer: []
      summary: Get all users method.
      tags:
      - users
    patch:
      consumes:
      - application/json
      description: Update user method.
      parameters:
      - description: User data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/user.UpdateUserInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.StatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - Bearer: []
      summary: Update user method.
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create user method.
      parameters:
      - description: User data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/user.CreateUserInput'
      produces:
      - application/json
      responses:
        "200":
          description: User ID
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - Bearer: []
      summary: Create user method.
      tags:
      - users
  /api/v1/users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete user method.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.StatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - Bearer: []
      summary: Delete user method.
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Get user by id method.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User data
          schema:
            $ref: '#/definitions/user.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - Bearer: []
      summary: Get user by id method.
      tags:
      - users
  /api/v1/users/roles/:
    get:
      consumes:
      - application/json
      description: Get all roles method.
      produces:
      - application/json
      responses:
        "200":
          description: Role List
          schema:
            items:
              $ref: '#/definitions/role.Role'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - Bearer: []
      summary: Get all roles method.
      tags:
      - roles
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"